fixed type

+++ src/cpu.hpp
@@ -376,7 +376,7 @@ namespace gbaemu
                 overflow = state.getFlag(cpsr_flags::V_FLAG_BITMASK),
                 carry = state.getFlag(cpsr_flags::C_FLAG_BITMASK);
 
-            uint32_t rnValue = state.accessReg(inst.params.block_data_transf.rn);
+            uint64_t rnValue = state.accessReg(inst.params.block_data_transf.rn);
             uint64_t resultValue;
 
             /* Different instructions cause different flags to be changed. */
